Update dev dependencies to latest versions
This commit updates the devDependencies in the package-lock.json file to
their latest versions. Specifically, it updates the following packages:
- @babel/preset-env from 7.23.5 to 7.23.3
- @babel/preset-react from 7.23.3 to 7.23.3
- babel-jest from 29.7.0 to 29.7.0
- jest from 27.5.1 to 27.5.1
- react-test-renderer from 18.2.0 to 18.2.0

By updating these devDependencies, we ensure that our development
environment is up to date and compatible with the latest versions of
these packages. This will allow us to take advantage of any bug fixes,
improvements, or new features introduced in these updates.

No issue references.